requirements.txt
transformers
accelerate
torch
langchain
langchain-openai
langchain-community
fastapi
uvicorn[standard]
python-multipart
Pillow
requests
python-dotenv
pydantic
openai
huggingface-hub
safetensors
tokenizers
numpy
tqdm
streamlit